requirements.txt
decord==0.6.0
einops==0.8.1
huggingface_hub==0.29.1
matplotlib==3.7.0
numpy==1.24.4
opencv_python==4.7.0.72
pyarrow==11.0.0
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.1.99
transformers==4.49.0
accelerate>=0.34.0
pytorch_lightning==2.6.0
nltk==3.9.2
rouge_score==0.1.2
radgraph==0.1.18
prdc==0.2
hi-ml-multimodal==0.2.2
wandb
pandas
timm
omegaconf
termcolor
torchmetrics[image]
gradio==3.47.1